Forward-port to sqlx 0.9 / rmcp 1.7 / pgvector 0.4.2; fix 4 NUMERIC decode bugs
An uncommitted `cargo update` had bumped pgvector 0.4.1->0.4.2 (which switched
to sqlx 0.9, creating a dual sqlx-core graph = 135 trait-mismatch errors) and
rmcp 1.1->1.7 (3 structs became #[non_exhaustive] = 3 errors). Forward-port onto
the new versions rather than reverting the lock.

rmcp 1.1 -> 1.7:
- AuthRequiredError / InsufficientScopeError via ::new() constructors.
- StreamableHttpServerConfig built via default() + the with_*() builder (the
  struct is now #[non_exhaustive]; the builder avoids field_reassign_with_default).
- Kept rmcp 1.7's secure loopback `allowed_hosts` default (DNS-rebinding
  protection): it matches the default mcp.host=127.0.0.1 so local/`/mcp` clients
  are unaffected; a remote deployment must call .with_allowed_hosts([...])
  (documented at the construction site).

sqlx 0.8 -> 0.9:
- query / query_as / query_scalar (+ _with) now take `impl SqlSafeStr`. Wrapped
  ~165 dynamic-SQL call sites in sqlx::AssertSqlSafe (audited: only structural
  tokens are interpolated -- column/table names, enum sql_in_list(), numeric
  limits -- every caller value stays a bind parameter).
- Multi-statement raw SQL in src/db/pool.rs -> sqlx::raw_sql(AssertSqlSafe(..)),
  which also fixes an E0521/E0716 'static lifetime in the after-connect closure.
- Manifests bumped in both Cargo.toml and pgmcp-testing/Cargo.toml (the test
  crate mirrors deps); single coherent sqlx 0.9 tree.

Also fixes 4 pre-existing NUMERIC -> float8/bigint decode mismatches that were
latent because the real-DB test suite is dormant on this host (the pgmcp role
lacks CREATEDB, so require_test_db! early-returns as "passed"). Found by running
CLI tools against the live DB; not migration regressions (sqlx 0.8 rejected them
too):
- db/queries/stats.rs total_bytes_indexed: SUM(bigint) widens to numeric.
- db/queries/metrics.rs, db/queries/concurrency.rs, digest/mod.rs: EXTRACT(EPOCH
  ...) returns numeric in PG14+.
Cast to ::float8 / ::bigint; live-verified (index_stats now reports real
counts/bytes; dependency_health returns rows).

verify.sh all gates green (build --all-targets, clippy -D warnings, release
suite, GPU smoke, TLA+/Rocq) on the forward-port; the decode fixes are
SQL-string-only and were re-verified by check/fmt/clippy/build --release plus a
live-DB re-run.

+++ src/digest/mod.rs
@@ -625,7 +625,7 @@ async fn collect_health(
 /// Whole days since a project was last scanned, or `None` when never scanned.
 async fn days_since_last_scan(pool: &PgPool, project_id: i32) -> Result<Option<i64>, sqlx::Error> {
     let days: Option<f64> = sqlx::query_scalar(
-        "SELECT EXTRACT(EPOCH FROM (now() - last_scanned_at)) / 86400.0
+        "SELECT (EXTRACT(EPOCH FROM (now() - last_scanned_at)) / 86400.0)::float8
            FROM projects WHERE id = $1 AND last_scanned_at IS NOT NULL",
     )
     .bind(project_id)
